Add Dockerfile and .dockerignore for worker service
- Introduced a Dockerfile implementing a multi-stage build for the worker service, optimizing the build process for production deployment.
- Created a .dockerignore file to exclude unnecessary files from the Docker build context, improving build efficiency.
- Updated ElasticsearchService to change the output field type from integer to long for better handling of large numbers.
- Enhanced docker-compose.yml by adding a Kibana service for monitoring Elasticsearch, including health checks and environment configurations.

+++ backend/services/worker/src/elasticsearch/elasticsearch.service.ts
@@ -70,7 +70,7 @@ export class ElasticsearchService implements OnModuleInit, OnModuleDestroy {
               jobId: { type: 'integer' },
               runId: { type: 'keyword' },
               input: { type: 'integer' },
-              output: { type: 'integer' },
+              output: { type: 'long' }, // Changed from integer to long for large numbers
               processedAt: { type: 'date' },
               timestamp: { type: 'date' },
             },
